+++ client/assets/components/document/document_jira/contentTypes/jiraProject.js
@@ -24,19 +24,32 @@
   }
 
 
-  function Controller(SignalsService) {
+  function Controller(SignalsService, PaginateService) {
     'ngInject';
     var vm = this;
 
     activate();
 
     function activate() {
-      vm.postSignal = SignalsService.postClickSignal;
+      vm.postSignal = postSignal;
       vm.doc = processDocument(vm.doc);
     }
 
     function processDocument(doc) {
+      doc.__signals_doc_id__ = SignalsService.getSignalsDocumentId(doc);
+      doc.page = PaginateService.getNormalizedCurrentPage();
       return doc;
+    }
+
+    function postSignal(options){
+      var paramsObj = {
+        params: {
+          position: vm.doc.position,
+          page: vm.doc.page
+        }
+      };
+      _.defaultsDeep(paramsObj, options);
+      SignalsService.postClickSignal(vm.doc.__signals_doc_id__, paramsObj);
     }
   }
 })();