+++ client/assets/components/document/document_web/document_web.js
@@ -15,6 +15,7 @@
       controllerAs: 'vm',
       bindToController: {
         doc: '=',
+        position: '=',
         highlight: '='
       }
     };
@@ -23,7 +24,7 @@
 
   }
 
-  function Controller(SignalsService) {
+  function Controller(SignalsService, PaginateService) {
     'ngInject';
     var vm = this;
 
@@ -31,7 +32,21 @@
 
 
     function activate() {
-      vm.postSignal = SignalsService.postClickSignal;
+      vm.postSignal = postSignal;
+      vm.doc.__signals_doc_id__ = SignalsService.getSignalsDocumentId(vm.doc);
+      vm.doc.position = vm.position;
+      vm.doc.page = PaginateService.getNormalizedCurrentPage();
+    }
+
+    function postSignal(options){
+      var paramsObj = {
+        params: {
+          position: vm.doc.position,
+          page: vm.doc.page
+        }
+      };
+      _.defaultsDeep(paramsObj, options);
+      SignalsService.postClickSignal(vm.doc.__signals_doc_id__, paramsObj);
     }
   }
 })();