+++ client/assets/js/services/services.js
@@ -8,5 +8,7 @@ angular.module('lucidworksView.services', [
   'lucidworksView.services.query',
   'lucidworksView.services.queryData',
   'lucidworksView.services.signals',
-  'lucidworksView.services.url'
+  'lucidworksView.services.url',
+  'lucidworksView.services.localParams',
+  'lucidworksView.services.clientStats'
 ]);